📝 Chore(#65): 코드 리뷰 반영
